Codes/requirements.txt
# Basic system and file handling
sys
os

# OpenCV for image processing
opencv-python>=4.5.0

# JSON handling
json

# Date and time manipulation
datetime

# PyTorch for deep learning and tensor operations
torch>=1.10.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Brian2 for computational neuroscience simulations
brian2>=2.4.0

# Optimization and mathematical functions
scipy.optimize

# Fast Fourier Transform (FFT) operations
scipy.fft

# Interpolation functions
scipy.interpolate

# Random number generation
random

# Pillow (PIL) for image processing
pillow>=8.0.0

# Matplotlib for data visualization
matplotlib>=3.4.0

# Matplotlib animation utilities
matplotlib.animation

# Advanced Matplotlib utilities
matplotlib.gridspec
matplotlib.ticker
mpl_toolkits.axes_grid1

# ScalarFormatter for scientific notation formatting
matplotlib.ticker.ScalarFormatter
